refactor: move function to datasets/util.py.
