+++ pytoda/datasets/_cache_datasource.py
@@ -5,15 +5,7 @@ import tempfile
 import warnings
 
 import diskcache as dc
-
-
-def sizeof_fmt(num, suffix='B'):
-    """Source: https://stackoverflow.com/a/1094933"""
-    for unit in ['', 'Ki', 'Mi', 'Gi', 'Ti', 'Pi', 'Ei', 'Zi']:
-        if abs(num) < 1024.0:
-            return "%3.1f%s%s" % (num, unit, suffix)
-        num /= 1024.0
-    return "%.1f%s%s" % (num, 'Yi', suffix)
+from .utils import sizeof_fmt
 
 
 class _CacheDatasource:
